+++ Utilities/ocvalidate/ValidateBooter.c
@@ -142,8 +142,6 @@ CheckBooterQuirks (
   UINT32                Index;
   OC_UEFI_DRIVER_ENTRY  *DriverEntry;
   CONST CHAR8           *Driver;
-  CONST CHAR8           *Load;
-  BOOLEAN               IsDriverLoaded;
   UINT8                 MaxSlide;
   BOOLEAN               IsAllowRelocationBlockEnabled;
   BOOLEAN               IsProvideCustomSlideEnabled;
@@ -166,19 +164,11 @@ CheckBooterQuirks (
   for (Index = 0; Index < Config->Uefi.Drivers.Count; ++Index) {
     DriverEntry    = Config->Uefi.Drivers.Values[Index];
     Driver         = OC_BLOB_GET (&DriverEntry->Path);
-    Load           = OC_BLOB_GET (&DriverEntry->Load);
-    IsDriverLoaded = (AsciiStrCmp (Load, "Early") == 0) || (AsciiStrCmp (Load, "Enabled") == 0);
 
     //
     // Skip sanitising UEFI->Drivers as it will be performed when checking UEFI section.
     //
-
-    if (!IsDriverLoaded && !(AsciiStrCmp (Load, "Disabled") == 0)) {
-      DEBUG ((DEBUG_WARN, "UEFI->Drivers[%d]->Load is borked (Can only be Early, Enabled, or Disabled)!\n", Index));
-      ++ErrorCount;
-    }
-
-    if (IsDriverLoaded && (AsciiStrCmp (Driver, "OpenRuntime.efi") == 0)) {
+    if (DriverEntry->Enabled && (AsciiStrCmp (Driver, "OpenRuntime.efi") == 0)) {
       HasOpenRuntimeEfiDriver = TRUE;
     }
   }